requirements.txt
streamlit
python-gitlab
python-dotenv
requests
gitlab-utils
pandas>=1.5
xlsxwriter
aiohttp
openpyxl
nest_asyncio
